backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9
pydantic>=2.7.1
scikit-learn>=1.5.0
numpy>=1.26.4
pandas>=2.2.2
joblib>=1.4.2
torch>=2.3.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
httpx>=0.27.0
pymongo[srv]>=4.7.2
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0